refactor: extract AgentActionCard component
Resolves #30618

Extracted the massive inline card component from UnifiedAgentFeed into a standalone reusable AgentActionCard component. This significantly improves maintainability of the UnifiedAgentFeed while maintaining all existing OHC premium tokens and minimum 44x44px touch targets. Verified functionality via E2E test suites.
