+++ src/ui/next/src/app/dashboard/UnifiedAgentFeed.tsx
@@ -5,6 +5,7 @@ import GrowthReferralWidget from "../components/GrowthReferralWidget";
 import { enqueueAction, getActions, removeAction } from "../utils/offlineQueue";
 import { AmbassadorReplyCard } from "./AmbassadorReplyCard";
 import { InstagramDMCard } from "./InstagramDMCard";
+import { AgentActionCard } from "../../components/feed/AgentActionCard";
 
 type TriageItem = {
   id: string;
@@ -129,7 +130,12 @@ export function UnifiedAgentFeed({ initialData }: { initialData?: any }) {
         const actions = await getActions();
         for (const action of actions) {
           if (action.type === "approve_agent_feed") {
-            await submitDecision(action.payload.id, action.payload.approved, action.payload.modified_content, action.payload.event_source);
+            await submitDecision(
+              action.payload.id,
+              action.payload.approved,
+              action.payload.modified_content,
+              action.payload.event_source,
+            );
             await removeAction(action.id);
             setOfflineActionsCount((prev) => Math.max(0, prev - 1));
             setQueuedActionIds((prev) => {
@@ -683,1578 +689,20 @@ export function UnifiedAgentFeed({ initialData }: { initialData?: any }) {
               </div>
             )}
             {items.map((approval) => (
-              <div
+              <AgentActionCard
                 key={approval.id}
-                className="glassmorphism bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] p-5 rounded-[16px] shadow-sm flex flex-col gap-4 transition-all duration-300"
-                data-testid={`triage-card-${approval.id}`}
-              >
-                <div className="flex flex-col gap-1">
-                  <div className="flex items-center gap-2">
-                    <span className="text-xs font-bold uppercase tracking-wider text-[#0066FF] bg-[#0066FF]/10 dark:bg-[#0066FF]/20 px-2 py-1 rounded-[8px]">
-                      Approval
-                    </span>
-                    <span className="text-xs font-bold uppercase tracking-wider text-indigo-600 dark:text-indigo-400 bg-indigo-50 dark:bg-indigo-900/30 px-2 py-1 rounded-[8px]">
-                      {approval.event_source.replace("_", " ")}
-                    </span>
-                    {approval.lifecycle_state === "PENDING_APPROVAL" && (
-                      <span className="text-xs font-bold uppercase tracking-wider text-green-700 bg-green-100 px-2 py-1 rounded-[8px]">
-                        Requires Review
-                      </span>
-                    )}
-                    {queuedActionIds.has(approval.id) && (
-                      <span
-                        className="text-xs font-bold uppercase tracking-wider text-yellow-600 bg-yellow-50 px-2 py-1 rounded-[8px] shadow-sm border border-yellow-200"
-                        data-testid="queued-badge"
-                      >
-                        Queued
-                      </span>
-                    )}
-                  </div>
-                  <h3 className="text-lg font-semibold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] leading-snug mt-1 tracking-wide">
-                    {approval.context_payload?.description ||
-                      approval.proposed_action?.message ||
-                      approval.proposed_action?.action_type ||
-                      approval.event_source}
-                  </h3>
-                  {((approval.proposed_action || approval.context_payload)
-                    ?.context ||
-                    (approval.proposed_action || approval.context_payload)
-                      ?.remaining_stock !== undefined ||
-                    (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "quote_draft" ||
-                    (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "social_post_draft" ||
-                    (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "ambassador_reply" ||
-                    (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "incident_resolution" ||
-                    (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "booking_draft" ||
-                    (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "instagram_dm") && (
-                    <div className="mt-2 flex flex-col gap-1 p-3 bg-gray-50 dark:bg-gray-800/50 rounded-[8px]">
-                      {(approval.proposed_action || approval.context_payload)
-                        ?.feature_type === "incident_resolution" && (
-                        <div
-                          className="mb-4 p-4 rounded-[16px] bg-red-50 dark:bg-red-900/20 border border-red-200 dark:border-red-800/50 flex flex-col gap-3"
-                          data-testid="incident-resolution-card"
-                        >
-                          <div className="flex items-center gap-2 text-red-600 font-semibold text-sm">
-                            <svg
-                              className="w-5 h-5"
-                              fill="none"
-                              stroke="currentColor"
-                              viewBox="0 0 24 24"
-                            >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth={2}
-                                d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-                              />
-                            </svg>
-                            CRITICAL INCIDENT
-                          </div>
-                          <p className="text-gray-700 dark:text-gray-300 text-sm break-words">
-                            {(
-                              approval.proposed_action ||
-                              approval.context_payload
-                            )?.description ||
-                              "An operational issue requires immediate attention."}
-                          </p>
-                        </div>
-                      )}
-                      {(approval.proposed_action || approval.context_payload)
-                        ?.feature_type === "onboarding_welcome" && (
-                        <div
-                          className="mb-4 p-4 rounded-[16px] bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] flex flex-col gap-3"
-                          data-testid="onboarding-welcome-card"
-                        >
-                          <div className="flex items-center gap-2 text-[#0066FF] font-semibold text-sm">
-                            <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" /></svg>
-                            Setup Complete
-                          </div>
-                          <p className="text-sm text-gray-800 dark:text-gray-200">
-                            {(approval.context_payload?.description || approval.proposed_action?.description) || "Welcome to OHC! I've set up your business. Click here to review your new storefront."}
-                          </p>
-                        </div>
-                      )}
-                      {(approval.proposed_action || approval.context_payload)
-                        ?.feature_type === "instagram_dm" && (
-                        <InstagramDMCard approval={approval} />
-                      )}
-                      {(approval.proposed_action || approval.context_payload)
-                        ?.feature_type === "ambassador_reply" && (
-                        <AmbassadorReplyCard approval={approval} />
-                      )}
-                      {(approval.proposed_action || approval.context_payload)
-                        ?.feature_type === "quote_draft" && (
-                        <div
-                          className="mb-4 p-4 rounded-[16px] bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] flex flex-col gap-3"
-                          data-testid="quote-draft-card"
-                        >
-                          <div className="flex items-center gap-2 text-[#0066FF] font-semibold text-sm">
-                            <svg
-                              className="w-5 h-5"
-                              fill="none"
-                              stroke="currentColor"
-                              viewBox="0 0 24 24"
-                            >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth={2}
-                                d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
-                              />
-                            </svg>
-                            Draft Quote:{" "}
-                            {(
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).service || "Plumbing Fix"}{" "}
-                            for Customer
-                          </div>
-                          <div className="text-xs text-[#0066FF] dark:text-blue-400 font-medium break-words">
-                            {
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).customer_inquiry
-                            }
-                          </div>
-                          <div className="bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] p-3 rounded-[8px] relative mt-2">
-                            <div className="text-[10px] uppercase font-bold text-gray-500 mb-2">
-                              AI Proposed Quote
-                            </div>
-                            <div className="space-y-2">
-                              <div className="flex justify-between">
-                                <span className="text-xs text-gray-500">
-                                  Calculated Total:
-                                </span>
-                                <span className="text-xs font-semibold text-gray-900 dark:text-gray-100">
-                                  $
-                                  {
-                                    (
-                                      approval.proposed_action ||
-                                      approval.context_payload
-                                    ).suggested_price
-                                  }
-                                </span>
-                              </div>
-                              <div className="flex justify-between">
-                                <span className="text-xs text-gray-500">
-                                  Scope of Work:
-                                </span>
-                                <span className="text-xs font-medium text-gray-800 dark:text-gray-200">
-                                  {
-                                    (
-                                      approval.proposed_action ||
-                                      approval.context_payload
-                                    ).scope
-                                  }
-                                </span>
-                              </div>
-                              <div className="flex justify-between">
-                                <span className="text-xs text-gray-500">
-                                  Suggested Time:
-                                </span>
-                                <span className="text-xs font-medium text-gray-800 dark:text-gray-200">
-                                  {
-                                    (
-                                      approval.proposed_action ||
-                                      approval.context_payload
-                                    ).suggested_time
-                                  }
-                                </span>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      )}
-                      {(approval.proposed_action || approval.context_payload)
-                        ?.feature_type === "newsletter_draft" ? (
-                        <div className="flex flex-col gap-3">
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400 font-semibold">
-                              Weekly Newsletter Draft Ready!
-                            </span>
-                            <span className="text-indigo-500 font-bold text-xs">
-                              Review and send
-                            </span>
-                          </div>
-
-                          <div className="bg-white/50 dark:bg-black/20 p-3 rounded-lg border border-gray-100 dark:border-gray-800">
-                            <div className="flex items-center gap-2 mb-2">
-                              <span className="text-xs font-semibold text-gray-500 uppercase tracking-wider">
-                                Subject
-                              </span>
-                            </div>
-                            <p className="text-sm font-medium text-gray-900 dark:text-gray-100 leading-snug">
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).subject
-                              }
-                            </p>
-                          </div>
-                          <div className="bg-white/50 dark:bg-black/20 p-3 rounded-lg border border-gray-100 dark:border-gray-800">
-                            <div className="flex items-center gap-2 mb-2">
-                              <span className="text-xs font-semibold text-gray-500 uppercase tracking-wider">
-                                Preview
-                              </span>
-                            </div>
-                            <p className="text-sm text-gray-700 dark:text-gray-300 italic line-clamp-3 leading-snug">
-                              "
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).content_preview
-                              }
-                              "
-                            </p>
-                          </div>
-                        </div>
-                      ) : (approval.proposed_action || approval.context_payload)
-                          ?.feature_type === "social_post_draft" ? (
-                        <div className="flex flex-col gap-3">
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400 font-semibold">
-                              New product detected!
-                            </span>
-                            <span className="text-pink-500 font-bold text-xs">
-                              Schedule a post?
-                            </span>
-                          </div>
-                          <div className="app-card dark:bg-gray-800 p-3 rounded-[8px] border border-pink-100 dark:border-pink-900/50">
-                            <div className="text-[10px] uppercase font-bold text-gray-400 mb-2 flex items-center gap-1">
-                              <span className="w-2 h-2 rounded-full bg-pink-500"></span>{" "}
-                              Instagram / TikTok Draft
-                            </div>
-                            <div className="text-xs text-gray-700 dark:text-gray-300 italic line-clamp-3">
-                              "
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).instagram ||
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).tiktok ||
-                                "Check out our new product!"}
-                              "
-                            </div>
-                          </div>
-                        </div>
-                      ) : (approval.proposed_action || approval.context_payload)
-                          ?.feature_type === "supply_order" ? (
-                        <>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Current Stock:
-                            </span>
-                            <span
-                              className="font-semibold text-gray-800 dark:text-gray-200"
-                              data-testid="supply-order-stock"
-                            >
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).remaining_stock
-                              }{" "}
-                              units
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Est. Runout:
-                            </span>
-                            <span className="font-semibold text-gray-800 dark:text-gray-200">
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).est_runout_days
-                              }{" "}
-                              days
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Reorder Quantity:
-                            </span>
-                            <span
-                              className="font-bold text-blue-600 dark:text-blue-400 text-base"
-                              data-testid="supply-order-quantity"
-                            >
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).suggested_reorder_quantity
-                              }{" "}
-                              Units
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm mb-3">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Vendor:
-                            </span>
-                            <span className="font-semibold text-gray-800 dark:text-gray-200">
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).vendor_name
-                              }{" "}
-                              (
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).vendor_contact
-                              }
-                              )
-                            </span>
-                          </div>
-                          <div className="bg-gray-50 dark:bg-gray-800 p-3 rounded-[8px] border border-gray-200 dark:border-gray-700">
-                            <div className="text-xs text-gray-500 dark:text-gray-400 mb-1">
-                              Drafted Message:
-                            </div>
-                            <div className="text-sm text-gray-800 dark:text-gray-200 italic font-medium">
-                              "
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).draft_message
-                              }
-                              "
-                            </div>
-                          </div>
-                        </>
-                      ) : (approval.proposed_action || approval.context_payload)
-                          ?.feature_type === "stockout_restock_and_price" ? (
-                        <>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Current Price:
-                            </span>
-                            <span className="font-semibold text-gray-400 dark:text-gray-500 line-through">
-                              $
-                              {Number(
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).old_price,
-                              ).toFixed(2)}
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Suggested Price:
-                            </span>
-                            <span
-                              className="font-bold text-green-600 dark:text-green-400 text-base"
-                              data-testid="stockout-new-price"
-                            >
-                              $
-                              {Number(
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).new_price,
-                              ).toFixed(2)}
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Reorder Quantity:
-                            </span>
-                            <span
-                              className="font-bold text-blue-600 dark:text-blue-400 text-base"
-                              data-testid="stockout-reorder"
-                            >
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).suggested_reorder_quantity
-                              }{" "}
-                              Units
-                            </span>
-                          </div>
-                          <div className="text-sm font-medium text-gray-800 dark:text-gray-200 mt-2">
-                            {
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).message
-                            }
-                          </div>
-                        </>
-                      ) : (approval.proposed_action || approval.context_payload)
-                          ?.context?.smart_pricing === true ? (
-                        <>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Current Price:
-                            </span>
-                            <span className="font-semibold text-gray-400 dark:text-gray-500 line-through">
-                              $
-                              {Number(
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).context.old_price,
-                              ).toFixed(2)}
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm mb-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Suggested Price:
-                            </span>
-                            <span
-                              className="font-bold text-green-600 dark:text-green-400 text-base"
-                              data-testid="smart-pricing-new-price"
-                            >
-                              $
-                              {Number(
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).context.new_price,
-                              ).toFixed(2)}
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Sales Projection:
-                            </span>
-                            <span
-                              className="font-semibold text-indigo-600 dark:text-indigo-400"
-                              data-testid="smart-pricing-sales-projection"
-                            >
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).context.sales_projection
-                              }
-                            </span>
-                          </div>
-                        </>
-                      ) : (approval.proposed_action || approval.context_payload)
-                          ?.feature_type === "incident_resolution" ? (
-                        <div className="flex flex-col gap-2">
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Proposed Actions:
-                            </span>
-                            <span className="font-semibold text-gray-900 dark:text-gray-100">
-                              {
-                                (
-                                  (
-                                    approval.proposed_action ||
-                                    approval.context_payload
-                                  ).actions || []
-                                ).length
-                              }{" "}
-                              steps
-                            </span>
-                          </div>
-                          <div className="w-full h-px bg-gray-200 dark:bg-gray-700 my-1"></div>
-                          {(
-                            (
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).actions || []
-                          ).map((action: any, idx: number) => (
-                            <div key={idx} className="flex flex-col mb-2">
-                              <span className="text-xs font-semibold text-gray-700 dark:text-gray-300">
-                                {action.action}
-                              </span>
-                              <span className="text-xs text-gray-500">
-                                {action.details}
-                              </span>
-                            </div>
-                          ))}
-                        </div>
-                      ) : (approval.proposed_action || approval.context_payload)
-                          ?.feature_type === "ambassador_reply" ? (
-                        <div className="flex flex-col gap-2">
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Context:
-                            </span>
-                            <span className="font-semibold text-gray-900 dark:text-gray-100">
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).source || "Message"}
-                            </span>
-                          </div>
-                          <div className="flex flex-col text-sm mt-1">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Draft:
-                            </span>
-                            <span className="font-semibold text-gray-900 dark:text-gray-100 line-clamp-2 mt-1">
-                              {
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).generated_response
-                              }
-                            </span>
-                          </div>
-                        </div>
-                      ) : (approval.proposed_action || approval.context_payload)
-                          ?.feature_type === "quote_draft" ? (
-                        <div className="flex flex-col gap-2">
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Context:
-                            </span>
-                            <span className="font-semibold text-gray-900 dark:text-gray-100">
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).customer_inquiry || "Client Inquiry"}
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Scope:
-                            </span>
-                            <span className="font-semibold text-gray-900 dark:text-gray-100">
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).scope ||
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).service}
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Timeline:
-                            </span>
-                            <span className="font-semibold text-gray-900 dark:text-gray-100">
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).suggested_time || "TBD"}
-                            </span>
-                          </div>
-                          <div className="flex justify-between items-center text-sm">
-                            <span className="text-gray-500 dark:text-gray-400">
-                              Price:
-                            </span>
-                            <span className="font-semibold text-green-600 dark:text-green-400">
-                              $
-                              {Number(
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                ).suggested_price ||
-                                  (
-                                    approval.proposed_action ||
-                                    approval.context_payload
-                                  ).price ||
-                                  0,
-                              ).toFixed(2)}
-                            </span>
-                          </div>
-                        </div>
-                      ) : (
-                        <>
-                          {(
-                            approval.proposed_action || approval.context_payload
-                          )?.context?.weekly_health_report === true ? (
-                            <div className="flex flex-col gap-2">
-                              <div className="text-sm text-gray-700 dark:text-gray-300">
-                                <span className="font-semibold">Summary:</span>{" "}
-                                {
-                                  (
-                                    approval.proposed_action ||
-                                    approval.context_payload
-                                  ).context.summary
-                                }
-                              </div>
-                              <div className="text-sm text-indigo-600 dark:text-indigo-400 font-medium">
-                                <span className="font-semibold text-gray-700 dark:text-gray-300">
-                                  Suggestion:
-                                </span>{" "}
-                                {
-                                  (
-                                    approval.proposed_action ||
-                                    approval.context_payload
-                                  ).context.actionable_suggestion
-                                }
-                              </div>
-                            </div>
-                          ) : (
-                            <>
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.context?.abandoned_carts_count !==
-                                undefined && (
-                                <div className="flex justify-between items-center text-sm">
-                                  <span className="text-gray-500 dark:text-gray-400">
-                                    Abandoned Carts:
-                                  </span>
-                                  <span className="font-semibold text-gray-900 dark:text-gray-100">
-                                    {
-                                      (
-                                        approval.proposed_action ||
-                                        approval.context_payload
-                                      ).context.abandoned_carts_count
-                                    }
-                                  </span>
-                                </div>
-                              )}
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.context?.potential_revenue !== undefined && (
-                                <div className="flex justify-between items-center text-sm">
-                                  <span className="text-gray-500 dark:text-gray-400">
-                                    Potential Revenue:
-                                  </span>
-                                  <span className="font-semibold text-green-600 dark:text-green-400">
-                                    $
-                                    {Number(
-                                      (
-                                        approval.proposed_action ||
-                                        approval.context_payload
-                                      ).context.potential_revenue,
-                                    ).toFixed(2)}
-                                  </span>
-                                </div>
-                              )}
-                              {(
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.remaining_stock !== undefined && (
-                                <div className="flex flex-col gap-2">
-                                  <div className="flex justify-between items-center text-sm">
-                                    <span className="text-gray-500 dark:text-gray-400">
-                                      Product ID:
-                                    </span>
-                                    <span className="font-semibold text-gray-900 dark:text-gray-100">
-                                      {
-                                        (
-                                          approval.proposed_action ||
-                                          approval.context_payload
-                                        ).product_id
-                                      }
-                                    </span>
-                                  </div>
-                                  <div className="flex justify-between items-center text-sm">
-                                    <span className="text-gray-500 dark:text-gray-400">
-                                      Remaining Stock:
-                                    </span>
-                                    <span className="font-semibold text-red-600 dark:text-red-400">
-                                      {
-                                        (
-                                          approval.proposed_action ||
-                                          approval.context_payload
-                                        ).remaining_stock
-                                      }
-                                    </span>
-                                  </div>
-                                  <div className="flex justify-between items-center text-sm">
-                                    <span className="text-gray-500 dark:text-gray-400">
-                                      Alert Message:
-                                    </span>
-                                    <span className="font-semibold text-gray-900 dark:text-gray-100">
-                                      {
-                                        (
-                                          approval.proposed_action ||
-                                          approval.context_payload
-                                        ).message
-                                      }
-                                    </span>
-                                  </div>
-                                </div>
-                              )}
-                            </>
-                          )}
-                        </>
-                      )}
-                    </div>
-                  )}
-                </div>
-
-                <div className="flex flex-col gap-3 w-full mt-2">
-                  {(approval.proposed_action || approval.context_payload)
-                    ?.feature_type === "onboarding_welcome" ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <a
-                        href="/storefront-builder"
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Review Storefront"
-                        data-testid="review-storefront-btn"
-                      >
-                        Review Storefront
-                      </a>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss"
-                        data-testid="dismiss-onboarding-welcome"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "booking_draft" ? (
-                    editingId === approval.id ? (
-                      <div className="flex flex-col gap-3 w-full">
-                        <textarea
-                          className="w-full min-h-[44px] p-3 rounded-[8px] border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-[#1D1D1F] dark:text-[#F5F5F7] text-sm focus:ring-2 focus:ring-[#0066FF] outline-none transition-all resize-none"
-                          rows={4}
-                          value={editContent}
-                          onChange={(e) => setEditContent(e.target.value)}
-                          data-testid="edit-booking-draft-textarea"
-                          autoFocus
-                        />
-                        <div className="flex gap-3">
-                          <button
-                            onClick={() => {
-                              handleDecision(
-                                approval.id,
-                                true,
-                                editContent,
-                                approval.event_source,
-                              );
-                              setEditingId(null);
-                            }}
-                            className="flex-1 min-h-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all shadow-md flex items-center justify-center"
-                            data-testid="save-booking-draft"
-                          >
-                            Save & Approve
-                          </button>
-                          <button
-                            onClick={() => setEditingId(null)}
-                            className="flex-1 min-h-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all flex items-center justify-center"
-                            data-testid="cancel-edit-booking-draft"
-                          >
-                            Cancel
-                          </button>
-                        </div>
-                      </div>
-                    ) : (
-                      <>
-                        <div className="p-3 bg-white/50 dark:bg-gray-800/50 rounded-lg mb-3 border border-gray-200 dark:border-gray-700 backdrop-blur-[10px]">
-                          <p className="text-sm font-medium text-gray-900 dark:text-white mb-1">
-                            Audio Summary
-                          </p>
-                          <div className="flex items-center gap-2 text-sm text-gray-600 dark:text-gray-400 bg-white dark:bg-gray-800 p-2 rounded border border-gray-200 dark:border-gray-700">
-                            <span className="text-lg">▶️</span>
-                            <span>0:10 AI Summary ({(approval.proposed_action || approval.context_payload)?.caller_phone})</span>
-                          </div>
-                        </div>
-                        <button
-                          onClick={() =>
-                            handleDecision(
-                              approval.id,
-                              true,
-                              undefined,
-                              approval.event_source,
-                            )
-                          }
-                          className="w-full min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all duration-200 shadow-md flex items-center justify-center mb-3"
-                          aria-label="Approve Route & Send Confirmation"
-                          data-testid="approve-booking-draft"
-                        >
-                          Approve Route & Send Confirmation
-                        </button>
-                        <div className="flex flex-col sm:flex-row gap-3 w-full">
-                          <button
-                            onClick={() => {
-                              setEditingId(approval.id);
-                              setEditContent(
-                                (approval.proposed_action || approval.context_payload)?.summary || ""
-                              );
-                            }}
-                            className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                            aria-label="Edit booking draft"
-                            data-testid="edit-booking-draft"
-                          >
-                            Edit
-                          </button>
-                          <button
-                            onClick={() =>
-                              handleDecision(
-                                approval.id,
-                                false,
-                                undefined,
-                                approval.event_source,
-                              )
-                            }
-                            className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                            aria-label="Reject booking draft"
-                            data-testid="reject-booking-draft"
-                          >
-                            Deny
-                          </button>
-                        </div>
-                      </>
-                    )
-                  ) : (approval.proposed_action || approval.context_payload)
-                    ?.feature_type === "incident_resolution" ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-red-600 text-white font-medium hover:bg-red-700 transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Execute Plan"
-                        data-testid="approve-incident-resolution"
-                      >
-                        Execute Plan
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss Plan"
-                        data-testid="dismiss-incident-resolution"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "instagram_dm" ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-pink-600 text-white font-medium hover:bg-pink-700 transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Approve & Send"
-                        data-testid="approve-instagram-dm"
-                      >
-                        Approve & Send
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss"
-                        data-testid="dismiss-instagram-dm"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "supply_order" ? (
-                    <>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="w-full min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-green-500 text-white font-medium hover:bg-green-600 transition-all duration-200 shadow-md flex items-center justify-center mb-3"
-                        aria-label="Approve & Send"
-                        data-testid="approve-supply-order"
-                      >
-                        Approve & Send
-                      </button>
-                      <div className="flex flex-col sm:flex-row gap-3 w-full">
-                        <button
-                          onClick={() => {
-                            setEditingId(approval.id);
-                            setEditContent(
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              ).draft_message,
-                            );
-                          }}
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Edit message"
-                          data-testid="edit-supply-order"
-                        >
-                          Edit
-                        </button>
-                        <button
-                          onClick={() =>
-                            handleDecision(
-                              approval.id,
-                              false,
-                              undefined,
-                              approval.event_source,
-                            )
-                          }
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Deny supply order"
-                          data-testid="reject-supply-order"
-                        >
-                          Deny
-                        </button>
-                      </div>
-                    </>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "newsletter_draft" ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-indigo-600 text-white font-medium hover:bg-indigo-700 transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Approve & Send"
-                      >
-                        Approve & Send
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Skip this week"
-                      >
-                        Skip this week
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "social_post_draft" ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-gradient-to-r from-pink-500 to-indigo-500 text-white font-medium hover:from-pink-600 hover:to-indigo-600 transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Schedule Posts"
-                        data-testid="approve-social-post"
-                      >
-                        Schedule Posts
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss proposal"
-                        data-testid="dismiss-social-post"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "supply_order" ? (
-                    <>
-                      <div className="flex justify-between items-center text-sm mb-1">
-                        <span className="text-gray-500 dark:text-gray-400">
-                          Current Stock:
-                        </span>
-                        <span
-                          className="font-semibold text-gray-800 dark:text-gray-200"
-                          data-testid="supply-order-stock"
-                        >
-                          {
-                            (
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).remaining_stock
-                          }{" "}
-                          units
-                        </span>
-                      </div>
-                      <div className="flex justify-between items-center text-sm mb-1">
-                        <span className="text-gray-500 dark:text-gray-400">
-                          Est. Runout:
-                        </span>
-                        <span className="font-semibold text-gray-800 dark:text-gray-200">
-                          {
-                            (
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).est_runout_days
-                          }{" "}
-                          days
-                        </span>
-                      </div>
-                      <div className="flex justify-between items-center text-sm mb-1">
-                        <span className="text-gray-500 dark:text-gray-400">
-                          Reorder Quantity:
-                        </span>
-                        <span
-                          className="font-bold text-blue-600 dark:text-blue-400 text-base"
-                          data-testid="supply-order-quantity"
-                        >
-                          {
-                            (
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).suggested_reorder_quantity
-                          }{" "}
-                          Units
-                        </span>
-                      </div>
-                      <div className="flex justify-between items-center text-sm mb-3">
-                        <span className="text-gray-500 dark:text-gray-400">
-                          Vendor:
-                        </span>
-                        <span className="font-semibold text-gray-800 dark:text-gray-200">
-                          {
-                            (
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).vendor_name
-                          }{" "}
-                          (
-                          {
-                            (
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).vendor_contact
-                          }
-                          )
-                        </span>
-                      </div>
-                      <div className="bg-gray-50 dark:bg-gray-800 p-3 rounded-[8px] border border-gray-200 dark:border-gray-700">
-                        <div className="text-xs text-gray-500 dark:text-gray-400 mb-1">
-                          Drafted Message:
-                        </div>
-                        <div className="text-sm text-gray-800 dark:text-gray-200 italic font-medium">
-                          "
-                          {
-                            (
-                              approval.proposed_action ||
-                              approval.context_payload
-                            ).draft_message
-                          }
-                          "
-                        </div>
-                      </div>
-                    </>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "stockout_restock_and_price" ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-green-600 text-white font-medium hover:bg-green-700 transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Approve"
-                        data-testid="approve-stockout"
-                      >
-                        Approve
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss"
-                        data-testid="dismiss-stockout"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "ambassador_reply" ? (
-                    editingId === approval.id ? (
-                      <div className="flex flex-col gap-3 w-full">
-                        <textarea
-                          className="w-full min-h-[44px] p-3 rounded-[8px] border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-[#1D1D1F] dark:text-[#F5F5F7] text-sm focus:ring-2 focus:ring-[#0066FF] outline-none transition-all resize-none"
-                          rows={4}
-                          value={editContent}
-                          onChange={(e) => setEditContent(e.target.value)}
-                          data-testid="edit-ambassador-reply-textarea"
-                          autoFocus
-                        />
-                        <div className="flex gap-3">
-                          <button
-                            onClick={() => {
-                              handleDecision(
-                                approval.id,
-                                true,
-                                editContent,
-                                approval.event_source,
-                              );
-                              setEditingId(null);
-                            }}
-                            className="flex-1 min-h-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all shadow-md flex items-center justify-center"
-                            data-testid="save-send-ambassador-reply"
-                          >
-                            Save & Send
-                          </button>
-                          <button
-                            onClick={() => setEditingId(null)}
-                            className="flex-1 min-h-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all flex items-center justify-center"
-                            data-testid="cancel-edit-ambassador-reply"
-                          >
-                            Cancel
-                          </button>
-                        </div>
-                      </div>
-                    ) : (
-                      <div className="flex flex-col sm:flex-row gap-3 w-full">
-                        <button
-                          onClick={() =>
-                            handleDecision(
-                              approval.id,
-                              true,
-                              undefined,
-                              approval.event_source,
-                            )
-                          }
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all duration-200 shadow-md flex items-center justify-center"
-                          aria-label="Approve & Send Draft"
-                          data-testid="approve-ambassador-reply"
-                        >
-                          Send Draft
-                        </button>
-                        <button
-                          onClick={() => {
-                            setEditingId(approval.id);
-                            setEditContent(
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.generated_response ||
-                                (
-                                  approval.proposed_action ||
-                                  approval.context_payload
-                                )?.draft_reply ||
-                                "",
-                            );
-                          }}
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Edit Draft"
-                          data-testid="edit-ambassador-reply"
-                        >
-                          Edit
-                        </button>
-                        <button
-                          onClick={() =>
-                            handleDecision(
-                              approval.id,
-                              false,
-                              undefined,
-                              approval.event_source,
-                            )
-                          }
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Dismiss Draft"
-                          data-testid="dismiss-ambassador-reply"
-                        >
-                          Dismiss
-                        </button>
-                      </div>
-                    )
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "quote_draft" ? (
-                    editingId === approval.id ? (
-                      <div className="flex flex-col gap-3 w-full">
-                        <div className="flex flex-col gap-1">
-                          <label className="text-xs text-gray-500 font-semibold">
-                            Total Price ($)
-                          </label>
-                          <input
-                            type="number"
-                            className="w-full p-3 rounded-[8px] border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-[#1D1D1F] dark:text-[#F5F5F7] text-sm focus:ring-2 focus:ring-[#0066FF] outline-none transition-all"
-                            value={editQuotePrice}
-                            onChange={(e) => setEditQuotePrice(e.target.value)}
-                            data-testid="edit-quote-price"
-                            autoFocus
-                          />
-                        </div>
-                        <div className="flex flex-col gap-1">
-                          <label className="text-xs text-gray-500 font-semibold">
-                            Scope of Work
-                          </label>
-                          <textarea
-                            className="w-full min-h-[44px] p-3 rounded-[8px] border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-[#1D1D1F] dark:text-[#F5F5F7] text-sm focus:ring-2 focus:ring-[#0066FF] outline-none transition-all resize-none"
-                            rows={3}
-                            value={editQuoteScope}
-                            onChange={(e) => setEditQuoteScope(e.target.value)}
-                            data-testid="edit-quote-scope"
-                          />
-                        </div>
-                        <div className="flex gap-3 mt-2">
-                          <button
-                            onClick={() => {
-                              handleDecision(
-                                approval.id,
-                                true,
-                                JSON.stringify({
-                                  price: parseFloat(editQuotePrice),
-                                  scope: editQuoteScope,
-                                }),
-                                approval.event_source,
-                              );
-                              setEditingId(null);
-                            }}
-                            className="flex-1 min-h-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all shadow-md flex items-center justify-center"
-                            data-testid="modal-approve-btn"
-                          >
-                            Approve & Send
-                          </button>
-                          <button
-                            onClick={() => setEditingId(null)}
-                            className="flex-1 min-h-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all flex items-center justify-center"
-                            data-testid="cancel-edit-quote"
-                          >
-                            Cancel
-                          </button>
-                        </div>
-                      </div>
-                    ) : (
-                      <div className="flex flex-col sm:flex-row gap-3 w-full">
-                        <button
-                          onClick={() =>
-                            handleDecision(
-                              approval.id,
-                              true,
-                              undefined,
-                              approval.event_source,
-                            )
-                          }
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all duration-200 shadow-md flex items-center justify-center"
-                          aria-label="Approve & Send"
-                          data-testid="approve-quote-draft"
-                        >
-                          Approve & Send
-                        </button>
-                        <button
-                          onClick={() => {
-                            setEditingId(approval.id);
-                            setEditQuotePrice(
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.suggested_price?.toString() || "",
-                            );
-                            setEditQuoteScope(
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.scope || "",
-                            );
-                          }}
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Edit Draft"
-                          data-testid="edit-quote-draft"
-                        >
-                          Edit
-                        </button>
-                      </div>
-                    )
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.context?.smart_pricing === true ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Approve & Run Sale"
-                        data-testid="approve-run-sale"
-                      >
-                        Approve & Run Sale
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss proposal"
-                        data-testid="dismiss-sale"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.context?.weekly_health_report === true ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-green-600 text-white font-medium hover:bg-green-700 transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Draft it"
-                        data-testid="approve-draft"
-                      >
-                        Yes, draft it!
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss proposal"
-                        data-testid="dismiss-draft"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.remaining_stock !== undefined ? (
-                    <div className="flex flex-col sm:flex-row gap-3 w-full">
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-amber-500 text-white font-medium hover:bg-amber-600 transition-all duration-200 shadow-md flex items-center justify-center"
-                        aria-label="Approve Restock"
-                        data-testid="approve-restock"
-                      >
-                        Approve Restock
-                      </button>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            false,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                        aria-label="Dismiss restock"
-                        data-testid="dismiss-restock"
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  ) : (approval.proposed_action || approval.context_payload)
-                      ?.feature_type === "quote_draft" ? (
-                    <>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="w-full min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-green-500 text-white font-medium hover:bg-green-600 transition-all duration-200 shadow-md flex items-center justify-center mb-3"
-                        aria-label="Approve & Send"
-                        data-testid="approve-send-proposal"
-                      >
-                        Approve & Send
-                      </button>
-                      <div className="flex flex-col sm:flex-row gap-3 w-full">
-                        <a
-                          href={`/quotes/${(approval.proposed_action || approval.context_payload)?.quote_id || approval.id}`}
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Edit Draft"
-                          data-testid="edit-proposal"
-                        >
-                          Edit Draft
-                        </a>
-                        <button
-                          onClick={() =>
-                            handleDecision(
-                              approval.id,
-                              false,
-                              undefined,
-                              approval.event_source,
-                            )
-                          }
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Ask Agent to Adjust"
-                          data-testid="reject-proposal"
-                        >
-                          Ask Agent to Adjust
-                        </button>
-                      </div>
-                    </>
-                  ) : editingId === approval.id ? (
-                    <div className="flex flex-col gap-3 w-full">
-                      <textarea
-                        className="w-full min-h-[44px] p-3 rounded-[8px] border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-[#1D1D1F] dark:text-[#F5F5F7] text-sm focus:ring-2 focus:ring-[#0066FF] outline-none transition-all resize-none"
-                        rows={4}
-                        value={editContent}
-                        onChange={(e) => setEditContent(e.target.value)}
-                        data-testid="edit-proposal-textarea"
-                        autoFocus
-                      />
-                      <div className="flex gap-3">
-                        <button
-                          onClick={() => {
-                            handleDecision(
-                              approval.id,
-                              true,
-                              editContent,
-                              approval.event_source,
-                            );
-                            setEditingId(null);
-                          }}
-                          className="flex-1 min-h-[44px] px-4 rounded-[8px] bg-[#0066FF] text-white font-medium hover:bg-[#0052CC] transition-all shadow-md flex items-center justify-center"
-                          data-testid="save-proposal"
-                        >
-                          Save & Approve
-                        </button>
-                        <button
-                          onClick={() => setEditingId(null)}
-                          className="flex-1 min-h-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all flex items-center justify-center"
-                          data-testid="cancel-edit-proposal"
-                        >
-                          Cancel
-                        </button>
-                      </div>
-                    </div>
-                  ) : (
-                    <>
-                      <button
-                        onClick={() =>
-                          handleDecision(
-                            approval.id,
-                            true,
-                            undefined,
-                            approval.event_source,
-                          )
-                        }
-                        className="w-full min-h-[44px] min-w-[44px] px-4 rounded-[8px] bg-green-500 text-white font-medium hover:bg-green-600 transition-all duration-200 shadow-md flex items-center justify-center mb-3"
-                        aria-label="Approve proposal"
-                        data-testid="approve-proposal"
-                      >
-                        Approve
-                      </button>
-                      <div className="flex flex-col sm:flex-row gap-3 w-full">
-                        <button
-                          onClick={() => {
-                            setEditingId(approval.id);
-                            const textToEdit =
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.generated_response ||
-                              (
-                                approval.proposed_action ||
-                                approval.context_payload
-                              )?.draft_reply ||
-                              approval.context_payload?.description ||
-                              approval.proposed_action?.message ||
-                              approval.proposed_action?.action_type ||
-                              approval.event_source;
-                            setEditContent(textToEdit || "");
-                          }}
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Edit proposal"
-                          data-testid="edit-proposal"
-                        >
-                          Edit
-                        </button>
-                        <button
-                          onClick={() =>
-                            handleDecision(
-                              approval.id,
-                              false,
-                              undefined,
-                              approval.event_source,
-                            )
-                          }
-                          className="flex-1 min-h-[44px] min-w-[44px] px-4 rounded-[8px] border border-gray-300 dark:border-gray-600 text-[#1D1D1F] dark:text-[#F5F5F7] font-medium hover:bg-gray-100 dark:hover:bg-gray-800 transition-all duration-200 flex items-center justify-center"
-                          aria-label="Reject proposal"
-                          data-testid={`triage-dismiss-${approval.id}`}
-                        >
-                          Deny
-                        </button>
-                      </div>
-                    </>
-                  )}
-                </div>
-              </div>
+                approval={approval}
+                queuedActionIds={queuedActionIds}
+                editingId={editingId}
+                editContent={editContent}
+                editQuotePrice={editQuotePrice}
+                editQuoteScope={editQuoteScope}
+                setEditingId={setEditingId}
+                setEditContent={setEditContent}
+                setEditQuotePrice={setEditQuotePrice}
+                setEditQuoteScope={setEditQuoteScope}
+                handleDecision={handleDecision}
+              />
             ))}
           </>
         )}
