+++ src/ui/next/test-results/.last-run.json
@@ -1,8 +1,6 @@
 {
   "status": "failed",
   "failedTests": [
-    "c39c44db5e2590c0a6f8-e67d23cc5f76decc3a05",
-    "c39c44db5e2590c0a6f8-bca6962e9ddb777a82a5",
-    "c39c44db5e2590c0a6f8-7775b5d2a15ce65939b2"
+    "a347544f07a30058c1ae-2d7f22ae8d19ae55e9c7"
   ]
 }
